requirements.txt
exa_py==1.13.1
firecrawl_py==2.7.0
langchain==0.3.25
langchain_astradb==0.6.0
langchain_core==0.3.62
langchain_google_genai==2.1.5
langgraph==0.4.7
langgraph_supervisor==0.0.26
pydantic==2.11.5
python-dotenv==1.1.0
tavily_python==0.7.0
